Split the integrations card into a component folder
One file defined eight JSX-returning functions and an icon registry; this PR
had grown it by 147 lines. The house rule is one component per file, and a file
made of several larger ones is split into a folder.

Nine files in `org-integrations/`: the card is now only the container — what
there is to show and how it is laid out — with everything it does moved into
`useOrgIntegrations`, and each piece of a row in its own file.

`renderProviderIcon` becomes a real `<ProviderIcon />`. That is the point of
`provider-icon.tsx`, not an incidental rename: a lowercase helper that returns
JSX is not a component and cannot be treated as one.

Every piece is a move. The characterization test passes unmodified — that is
the proof — including the stuck-button quirk and the untranslated empty state,
both of which are deliberately left exactly as they were.

Verified: pnpm check 32/32; vitest 179 files, 2656 tests, unchanged.
