+++ apps/app/src/features/integrations/client.ts
@@ -1,3 +1,3 @@
 "use client";
 
-export { OrgIntegrationsCard } from "./settings/components/org-integrations-card";
+export { OrgIntegrationsCard } from "./settings/components/org-integrations/org-integrations-card";
