+++ apps/app/src/features/integrations/settings/components/org-integrations/provider-grants.tsx
@@ -1,0 +1,72 @@
+"use client";
+
+import type { IntegrationProviderId } from "@/features/integrations/contracts";
+import type { OrgSettingsPage } from "@/features/organizations/contracts";
+
+import { ExternalLink } from "lucide-react";
+import { useEffect } from "react";
+import { toast } from "sonner";
+
+import { api } from "@/shared/api/trpc/client";
+
+// Its own query, so the card renders at once and only this strip waits on the
+// provider.
+export const ProviderGrants = ({
+  orgSlug,
+  provider,
+  t,
+}: {
+  orgSlug: string;
+  provider: IntegrationProviderId;
+  t: OrgSettingsPage["integrations"];
+}) => {
+  const utils = api.useUtils();
+  const { data, isPending, isError, error } =
+    api.integration.listGrants.useQuery({
+      orgSlug,
+      provider,
+    });
+
+  // The token this query needs is minted per call, so this strip is where an
+  // uninstall on the provider's side first shows up. The server has already
+  // dropped the connection by the time the error arrives; refetching the list
+  // is what takes the row off the page.
+  const wasRevoked = error?.data?.applicationCode === "integration.revoked";
+  useEffect(() => {
+    if (!wasRevoked) return;
+    toast.error(t.revokedNotice, { id: `integration-revoked-${provider}` });
+    void utils.integration.list.invalidate({ orgSlug });
+  }, [wasRevoked, provider, orgSlug, t.revokedNotice, utils]);
+
+  if (isPending) {
+    return <p className="text-[11px] text-neutral-400">{t.grantsLoading}</p>;
+  }
+  if (isError) {
+    return <p className="text-[11px] text-neutral-400">{t.grantsError}</p>;
+  }
+  if (data.grants.length === 0) {
+    return <p className="text-[11px] text-neutral-400">{t.grantsEmpty}</p>;
+  }
+  return (
+    <div className="flex flex-col gap-1">
+      <p className="text-[11px] font-medium text-neutral-500 dark:text-neutral-400">
+        {t.grantsTitle}
+      </p>
+      <ul className="flex flex-wrap gap-1.5">
+        {data.grants.map((grant) => (
+          <li key={grant.id}>
+            <a
+              href={grant.url}
+              target="_blank"
+              rel="noreferrer"
+              className="inline-flex items-center gap-1 rounded-md border border-neutral-200 px-2 py-0.5 text-[11px] text-neutral-600 hover:bg-neutral-50 dark:border-neutral-700 dark:text-neutral-300 dark:hover:bg-neutral-800"
+            >
+              {grant.name}
+              <ExternalLink className="h-2.5 w-2.5" />
+            </a>
+          </li>
+        ))}
+      </ul>
+    </div>
+  );
+};
